Default cache validation to digest mode in CI
A cache restored across CI runs never passes ADR-87's stat-tier
freshness check: a fresh checkout regenerates every size/mtime/ctime/
inode tuple. File entries survive through their content-digest
fallback, but plugin watch-glob slots carry a stat-only signature and
recomputed their producers on every run — established empirically in
PR #189, which worked around it by exporting RIGOR_STRICT_VALIDATION
in our own workflows and documenting the trap.

Resolve it at the source instead. `cache.validation` gains an `auto`
value, now the default: strict (content-digest) when the ADR-51
CiDetector recognises a CI provider, stat-first everywhere else.
Explicit `stat` / `digest` win outright — `stat` is the opt-out for a
self-hosted runner with a persistent workspace, where the tuples are
stable and the null-build floor is worth keeping. RIGOR_CI_DETECT=0
disables detection (the seam the spec suite already sets globally, so
suite determinism is unchanged even on GitHub Actions), and
RIGOR_STRICT_VALIDATION=1 keeps winning over everything. Resolution
happens per run rather than at Configuration construction so the
frozen carrier stays environment-independent.

CiDetector moves out of the CLI namespace to Rigor::CiDetector: it is
a pure function of an env hash with no CLI dependency, and
Configuration must not require upward into cli/.

Validation mode never changes analysis results — only where the
freshness cost lands — so the flip is FP-safe by construction. A cache
recorded under the opposite mode simply mismatches and re-records
once. Our workflows drop the now-redundant env var and dogfood the
auto path.

Closes #190.
