+++ docs/manual/11-ci.md
@@ -254,18 +254,14 @@ deterministic, and fast enough for most projects. When the Rigor
 job's runtime starts to matter (a large codebase, many plugins),
 persist [the analysis cache](12-caching.md) between runs.
 
-Two things make this work:
-
-1. **Restore and save the cache directory** with your CI's cache
-   facility, keyed on anything that changes the analysis wholesale
-   (the Rigor version / lockfile).
-2. **Set `RIGOR_STRICT_VALIDATION=1`.** A fresh checkout
-   regenerates every file's timestamps and inode, so Rigor's
-   default stat-based freshness check can never short-circuit in CI,
-   and plugin watch-glob entries — which are validated by stat
-   signature alone — would be recomputed on every run. Strict mode
-   validates by content hash, which is identical across checkouts.
-   (Details: [How a file is checked for changes](12-caching.md#how-a-file-is-checked-for-changes).)
+**Restore and save the cache directory** with your CI's cache
+facility, keyed on anything that changes the analysis wholesale
+(the Rigor version / lockfile). That is the whole setup: Rigor
+detects CI and switches its cache freshness check from stat
+metadata (which a fresh checkout regenerates, and which would
+otherwise force parts of the restored cache to recompute every
+run) to content hashes, which are identical across checkouts.
+Details: [How a file is checked for changes](12-caching.md#how-a-file-is-checked-for-changes).
 
 GitHub Actions:
 
@@ -276,21 +272,25 @@ GitHub Actions:
           key: rigor-cache-${{ runner.os }}-${{ hashFiles('.github/rigor/Gemfile.lock') }}-${{ github.sha }}
           restore-keys: rigor-cache-${{ runner.os }}-${{ hashFiles('.github/rigor/Gemfile.lock') }}-
       - run: bundle exec rigor check
-        env:
-          RIGOR_STRICT_VALIDATION: "1"
 ```
 
 GitLab CI:
 
 ```yaml
 rigor:
-  variables:
-    RIGOR_STRICT_VALIDATION: "1"
   cache:
     key: rigor-cache
     paths:
       - .rigor/cache
 ```
+
+Two refinements:
+
+- A **self-hosted runner that reuses its workspace** keeps stable
+  stat metadata, so the faster stat check works there — opt back
+  into it with `cache.validation: stat`.
+- On Rigor versions before the `auto` default, set
+  `RIGOR_STRICT_VALIDATION: "1"` on the job explicitly.
 
 The cache is content-keyed and self-invalidating — a stale or
 corrupt restore degrades to a recompute, never a wrong result. If
