+++ spec/rigor/cli_spec.rb
@@ -1494,7 +1494,7 @@ RSpec.describe Rigor::CLI do
     # All known CI provider variables are also saved and cleared so that running the suite under a real CI (e.g. GitHub
     # Actions) does not bleed its own GITHUB_ACTIONS=true into tests that simulate a different platform.
     def with_ci_env(vars)
-      all_provider_vars = Rigor::CLI::CiDetector::PROVIDERS.map { |p| p[:var] }
+      all_provider_vars = Rigor::CiDetector::PROVIDERS.map { |p| p[:var] }
       keys = (vars.keys + ["RIGOR_CI_DETECT"] + all_provider_vars).uniq
       saved = keys.to_h { |k| [k, ENV.fetch(k, nil)] }
       keys.each { |k| ENV.delete(k) }
